Move auth to lib directory
